Add support for px unit stripping. (#5)
I ran into an issue using postcss-pr along with
postcss-vertical-rhythm. The latter outputs px units,
but the former expects a float value without a unit. This
change allows for the use of postcss-vertical-rhythm's property
value syntax, such as "1vr", to be combined with the "pr"
unit - e.g., margin: (1vr)pr - and successfully
output into the "rem" equivalent.